--- timecopilot/models/benchmarks/__init__.py
@@ -1,29 +1,0 @@
-from .prophet import Prophet
-from .stats import (
-    ADIDA,
-    IMAPA,
-    AutoARIMA,
-    AutoCES,
-    AutoETS,
-    CrostonClassic,
-    DynamicOptimizedTheta,
-    HistoricAverage,
-    SeasonalNaive,
-    Theta,
-    ZeroModel,
-)
-
-__all__ = [
-    "ADIDA",
-    "AutoARIMA",
-    "AutoCES",
-    "AutoETS",
-    "CrostonClassic",
-    "DynamicOptimizedTheta",
-    "HistoricAverage",
-    "Prophet",
-    "IMAPA",
-    "SeasonalNaive",
-    "Theta",
-    "ZeroModel",
-]
