+++ timecopilot/models/foundation/tirex.py
