+++ timecopilot/models/foundation/utils.py
